uki: split out os-release code to separate file
We'll want to use this for non-UKI cases soon, as well.
